requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.6
    # via py-cord
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
charset-normalizer==3.3.2
    # via aiohttp
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
py-cord==2.4.1
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
typing-extensions==4.8.0
    # via py-cord
yarl==1.9.3
    # via aiohttp
